fixed issue 273 ?
git-svn-id: http://tint2.googlecode.com/svn/trunk@518 121b4492-b84c-0410-8b4c-0d4edfb3f3cc

+++ src/panel.c
@@ -495,7 +495,7 @@ void set_panel_properties(Panel *p)
 	XChangeProperty(server.dsp, p->main_win, server.atom._MOTIF_WM_HINTS, server.atom._MOTIF_WM_HINTS, 32, PropModeReplace, (unsigned char *) prop, 5);
 
 	// XdndAware - Register for Xdnd events
-	int version=5;
+	long version=5;
 	XChangeProperty(server.dsp, p->main_win, server.atom.XdndAware, XA_ATOM, 32, PropModeReplace, (unsigned char*)&version, 1);
 
 	update_strut(p);
